move internal a11y stories to the right directory
